Added calculation of lever arm and K factor.

+++ src/main/java/com/radsoltan/model/geometry/Shape.java
@@ -7,6 +7,6 @@ public abstract class Shape implements Drawable {
     public abstract double calculateArea();
     public abstract double calculateFirstMomentOfArea();
     public abstract double calculateSecondMomentOfArea();
-    public abstract double getWidthInCompressionZone(double UlsMoment, double effectiveDepth, double fcd);
+    public abstract int getWidthInCompressionZone(double UlsMoment, double effectiveDepth, double fcd);
     public abstract int getDepth();
 }
